+++ packages/plugin-stylelint/src/index.ts
@@ -1,7 +1,1 @@
-export {
-  stylelint,
-  stylelintExtensions,
-  stylelintFlags,
-  stylelintIgnorePatterns,
-} from './plugin-stylelint';
-export type {StylelintFlags} from './plugin-stylelint';
+export {stylelint} from './plugin-stylelint';